Issue: #8 게임 상태 로직을 개별 클래스로 분리
