+++ Assets/Scripts/Game/BuildState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b3eba016bb4b7a546a2525e22aa32c44
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
