+++ Assets/Scripts/Game/MenuState.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MenuState : MonoBehaviour
+{
+    private void Start()
+    {
+        
+    }
+
+    private void Update()
+    {
+        
+    }
+}
