+++ Assets/Scripts/Game/MenuState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 446ce159c6688214686409b73b4de9f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
