+++ Assets/Scripts/Game/PlayState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5974b1e53ae788243bdb51c7f7bcacf8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
